requirements.txt
kafka-python
pandas
boto3
gzip
